renames to match structure
