+++ source/extensions/omni.isaac.lab/omni/isaac/lab/envs/manager_based_env_cfg.py
@@ -82,6 +82,7 @@ class ManagerBasedEnvCfg:
     # simulation settings
     viewer: ViewerCfg = ViewerCfg()
     """Viewer configuration. Default is ViewerCfg()."""
+
     sim: SimulationCfg = SimulationCfg()
     """Physics simulation configuration. Default is SimulationCfg()."""
